+++ x/vpool/keeper/pool_state.go
@@ -29,7 +29,7 @@ func (k Keeper) CreatePool(
 	)
 
 	k.savePool(ctx, pool)
-	k.saveSnapshot(ctx, pair, 0, pool.QuoteAssetReserve, pool.BaseAssetReserve, ctx.BlockTime(), ctx.BlockHeight())
+	k.saveSnapshot(ctx, pair, 0, pool.QuoteAssetReserve, pool.BaseAssetReserve)
 	k.saveSnapshotCounter(ctx, pair, 0)
 }
 
@@ -61,12 +61,12 @@ Saves an updated pool to state and snapshots it.
 args:
   - ctx: cosmos-sdk context
   - updatedPool: pool object to save to state
-  - skipFluctuationCheck: override fluctuation check from last snapshot
+  - skipFluctuationCheck: determines if a fluctuation check should be done against the last snapshot
 
 ret:
   - err: error
 */
-func (k Keeper) savePoolAndSnapshot(
+func (k Keeper) updatePool(
 	ctx sdk.Context,
 	updatedPool *types.Pool,
 	skipFluctuationCheck bool,
@@ -78,7 +78,7 @@ func (k Keeper) savePoolAndSnapshot(
 		}
 	}
 
-	if err = k.addReserveSnapshot(
+	if err = k.updateSnapshot(
 		ctx,
 		updatedPool.Pair,
 		updatedPool.QuoteAssetReserve,